Make Request component less unintuitive
shouldComponentUpdate was implemented to avoid an initial double render
due to setting the state in componentDidMount.
The trade off was that the render function passed as children needed to be
pure, and it could lead to unintuitive behaviour if it was not (for example
if it was a method that used this to render its contents).
I don't think the trade off was worth it, as the double render happens
for the loading state, which should typically be a decently fast render.